shiina

+++ ssm01/src/main/java/com/itcast/service/impl/UserServiceImpl.java
@@ -32,7 +32,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public boolean insertUser(String username, String money) {
-        long i = userMapper.insertUser(username, money);
+        Integer i = userMapper.insertUser(username, money);
         System.out.println("service...");
         return i>0;
     }
